Move compose.* tests to own namespace
